Instead of copying the SelfDeployUI project next to the SelfDeploy.dll
Using a PreBuild event it runs PowerShell script to zip the SelfDeployUI project as UI.zip
the UI.zip is an embedded resource in the SelfDeploy project
then at run time we extract the project to a temp directory.  We upload the zip file of the directory into Archive.Zip
we then run dotnet build command against SelfDeployUI  temp folder project
We pick up the EXE that was generated can copy to the end result path
We delete the temp directory

While debugging this found that we had to include the System.Resources.Extensions if outside of Visual Studio to build... !??! or run MSBUILD  Since MSBuild is tougher to find... stuck to adding Nuget Package.

No need in the future commit the UI.zip file as it will change but that is cause it happens on every build.